Rename directory as it used Cypress terminology
