+++ frontend/benefit/handler/browser-tests/json/list-handling.json
