+++ frontend/benefit/handler/browser-tests/json/list-received.json
